+++ src/qn/qnitensor.jl
@@ -88,24 +88,8 @@ function ITensor(::UndefInitializer,
   return ITensor(Float64,undef,flux,IndexSet(inds...))
 end
 
-
-function _ITensor(A::Array{ElT}, inds::QNIndexSet; tol = 0) where {ElT}
-  length(A) ≠ dim(inds) && throw(DimensionMismatch("In ITensor(::Array, ::IndexSet), length of Array ($(length(A))) must match total dimension of IndexSet ($(dim(inds)))"))
-  T = emptyITensor(ElT, inds)
-  A = reshape(A, dims(inds))
-  for vs in eachindex(T)
-    Avs = A[vs]
-    if abs(Avs) > tol
-      T[vs] = A[vs]
-    end
-  end
-  return T
-end
-
 """
-    ITensor(::Array, ::IndexSet; tol = 0)
-
-    ITensor(::Array, ::Index...; tol = 0)
+    ITensor([ElT::Type, ]::Array, inds; tol = 0)
 
 Create a block sparse ITensor from the input Array, and collection 
 of QN indices. Zeros are dropped and nonzero blocks are determined
@@ -115,6 +99,7 @@ Optionally, you can set a tolerance such that elements
 less than or equal to the tolerance are dropped.
 
 # Examples
+
 ```julia
 julia> i = Index([QN(0)=>1, QN(1)=>2], "i");
 
@@ -138,37 +123,23 @@ Block: (2, 2)
  0.0  4.0
 ```
 """
-ITensor(A::Array, inds::QNIndexSet; tol = 0) =
-  _ITensor(A, inds; tol = tol)
+function itensor(::Type{ElT}, A::Array{<: Number}, inds::QNIndexSet; tol = 0) where {ElT <: Number}
+  length(A) ≠ dim(inds) && throw(DimensionMismatch("In ITensor(::Array, ::IndexSet), length of Array ($(length(A))) must match total dimension of IndexSet ($(dim(inds)))"))
+  T = emptyITensor(ElT, inds)
+  A = reshape(A, dims(inds))
+  for vs in eachindex(T)
+    Avs = A[vs]
+    if abs(Avs) > tol
+      T[vs] = A[vs]
+    end
+  end
+  return T
+end
 
-# Defined to fix ambiguity error
-ITensor(A::Array{<: AbstractFloat}, inds::QNIndexSet; tol = 0) =
-  _ITensor(A, inds; tol = tol)
-
-# Defined to fix ambiguity error
-ITensor(A::Array{<: RealOrComplex{Int}}, inds::QNIndexSet; tol = 0) =
-  _ITensor(A, inds; tol = tol)
-
-ITensor(A::Array, inds::QNIndex...; tol = 0) =
-  _ITensor(A, IndexSet(inds...); tol = tol)
-
-# Defined to fix ambiguity error
-ITensor(A::Array{<: RealOrComplex{Int}}, inds::QNIndex...; tol = 0) =
-  _ITensor(A, IndexSet(inds...); tol = tol)
-
-itensor(A::Array, inds::QNIndexSet; tol = 0) =
-  ITensor(A, inds; tol = tol)
-
-itensor(A::Array, inds::QNIndex...; tol = 0) =
-  ITensor(A, inds...; tol = tol)
-
-# Defined to fix ambiguity error
-itensor(A::Array{<: Number}, inds::QNIndex...; tol = 0) =
-  ITensor(A, inds...; tol = tol)
-
-# Defined to fix ambiguity error
-itensor(A::Array{<: RealOrComplex{Int}}, inds::QNIndex...; tol = 0) =
-  ITensor(A, inds...; tol = tol)
+# Short-circuit the non-QN version that does a copy here
+function ITensor(::Type{ElT}, A::Array{ElT}, inds::QNIndexSet; kwargs...) where {ElT}
+  return itensor(ElT, A, inds; kwargs...)
+end
 
 """
     emptyITensor([::Type{ElT} = Float64, ]inds)